requirements.txt
py-cord >= 2.0.0
pyyaml >= 5.4.1
apscheduler == 3.8.0
mysql-connector-python >= 8.1.0